Moved things from client to server
Added getCurrentDateTime for naming curl files.

handleClient function added changes that if compromise over https in line then it sends a message to download using curl the file on the client and then transfers the file back to the server. The function downloadFile is where the file is actually recieved from the client.

searchStrings is in pySysLib that finds strings that might be compromises which is server side.

encryptData and decryptData are in pySysLib.